Refactor common/ directory to consolidate logic applicable to types.go
