--- common/big.go
@@ -1,14 +1,0 @@
-package common
-
-import "math/big"
-
-// Common big integers often used
-var (
-	Big1   = big.NewInt(1)
-	Big2   = big.NewInt(2)
-	Big3   = big.NewInt(3)
-	Big0   = big.NewInt(0)
-	Big32  = big.NewInt(32)
-	Big256 = big.NewInt(256)
-	Big257 = big.NewInt(257)
-)
